Changing hal.device.query.i32 -> hal.device.query.i64. Now that we use i64 values for things it makes more sense to do this and then truncate to i32 when needed. This will allow us to return things like maximum semaphore values, maximum allocation sizes, free memory, etc.

+++ compiler/src/iree/compiler/Dialect/HAL/hal.imports.mlir
@@ -300,11 +300,11 @@ vm.import @device.allocator(
 attributes {nosideeffects}
 
 // Returns a tuple of (ok, value) for the given configuration key.
-vm.import @device.query.i32(
+vm.import @device.query.i64(
   %device : !vm.ref<!hal.device>,
   %category : !vm.buffer,
   %key : !vm.buffer
-) -> (i32, i32)
+) -> (i32, i64)
 attributes {nosideeffects}
 
 //===----------------------------------------------------------------------===//
